# internal/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CreateWorkspaceRequest:
    properties:
      name:
        type: string
      package_manager:
        type: string
      path:
        type: string
      pixi_toml:
        type: string
      source:
        type: string
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.Workspace:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      owner:
        $ref: '#/definitions/models.User'
      owner_id:
        type: string
      package_manager:
        description: '"pixi" or "uv"'
        type: string
      path:
        description: filesystem path (local-mode)
        type: string
      size_bytes:
        type: integer
      source:
        description: '"managed", "local"'
        type: string
      status:
        $ref: '#/definitions/models.WorkspaceStatus'
      updated_at:
        type: string
    type: object
  models.WorkspaceStatus:
    enum:
    - pending
    - creating
    - ready
    - failed
    - deleting
    type: string
    x-enum-varnames:
    - WsStatusPending
    - WsStatusCreating
    - WsStatusReady
    - WsStatusFailed
    - WsStatusDeleting
...
host: localhost:8460
info:
  contact: {}
  description: Multi-User Workspace Management System API
  title: Nebi API
  version: "1.0"
paths:
...
  /workspaces:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Workspace'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all workspaces for the current user
      tags:
      - workspaces
    post:
      consumes:
      - application/json
      parameters:
      - description: Workspace details
        in: body
        name: workspace
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateWorkspaceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Workspace'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new workspace
      tags:
      - workspaces
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"